fix: UKFはf32で実行する

+++ src/ukf.rs
@@ -1,28 +1,28 @@
 extern crate nalgebra as na;
 
 pub struct UnscentedKalmanFilter {
-    x: na::Vector4<f64>,
-    p: na::Matrix4<f64>,
-    q: na::Matrix4<f64>,
-    r: na::Matrix2<f64>,
-    wm: na::SVector<f64, 9>,
-    wc: na::SVector<f64, 9>,
-    sigma_f: na::SMatrix<f64, 4, 9>,
+    x: na::Vector4<f32>,
+    p: na::Matrix4<f32>,
+    q: na::Matrix4<f32>,
+    r: na::Matrix2<f32>,
+    wm: na::SVector<f32, 9>,
+    wc: na::SVector<f32, 9>,
+    sigma_f: na::SMatrix<f32, 4, 9>,
 }
 
 impl UnscentedKalmanFilter {
-    const N: f64 = 4.0;
-    const ALPHA: f64 = 1e-3;
-    const BETA: f64 = 2.0;
-    const KAPPA: f64 = 3.0 - Self::N;
-    const C: f64 = Self::ALPHA * Self::ALPHA * (Self::N + Self::KAPPA);
-    const LAMBDA: f64 = Self::C - Self::N;
+    const N: f32 = 4.0;
+    const ALPHA: f32 = 1e-3;
+    const BETA: f32 = 2.0;
+    const KAPPA: f32 = 3.0 - Self::N;
+    const C: f32 = Self::ALPHA * Self::ALPHA * (Self::N + Self::KAPPA);
+    const LAMBDA: f32 = Self::C - Self::N;
 
     pub fn new(
-        x: na::Vector4<f64>,
-        p: na::Matrix4<f64>,
-        q: na::Matrix4<f64>,
-        r: na::Matrix2<f64>,
+        x: na::Vector4<f32>,
+        p: na::Matrix4<f32>,
+        q: na::Matrix4<f32>,
+        r: na::Matrix2<f32>,
     ) -> Self {
         let (wm, wc) = Self::sigma_weight();
         let sigma_f = Self::compute_sigma_points(x, p, 0.0, |x, _| x);
@@ -37,7 +37,7 @@ impl UnscentedKalmanFilter {
         }
     }
 
-    pub fn predict(&mut self, u: f64, fx: fn(na::Vector4<f64>, f64) -> na::Vector4<f64>) {
+    pub fn predict(&mut self, u: f32, fx: fn(na::Vector4<f32>, f32) -> na::Vector4<f32>) {
         self.sigma_f = Self::compute_sigma_points(self.x, self.p, u, fx);
         let (x, p) = Self::unscented_transform(&self.sigma_f, &self.wm, &self.wc, &self.q);
         self.x = x;
@@ -46,15 +46,15 @@ impl UnscentedKalmanFilter {
 
     pub fn update(
         &mut self,
-        x_obs: &na::Vector2<f64>,
-        hx: fn(na::Vector4<f64>) -> na::Vector2<f64>,
+        x_obs: &na::Vector2<f32>,
+        hx: fn(na::Vector4<f32>) -> na::Vector2<f32>,
     ) {
-        let mut sigmas_h = na::SMatrix::<f64, 2, 9>::zeros();
+        let mut sigmas_h = na::SMatrix::<f32, 2, 9>::zeros();
         for i in 0..9 {
             sigmas_h.set_column(i, &hx(self.sigma_f.column(i).into_owned()));
         }
         let (zp, pz) = Self::unscented_transform(&sigmas_h, &self.wm, &self.wc, &self.r);
-        let mut pxz = na::SMatrix::<f64, 4, 2>::zeros();
+        let mut pxz = na::SMatrix::<f32, 4, 2>::zeros();
         for i in 0..9 {
             pxz += self.wc[i]
                 * (self.sigma_f.column(i) - self.x)
@@ -68,11 +68,11 @@ impl UnscentedKalmanFilter {
     }
 
     fn compute_sigma_points(
-        x: na::Vector4<f64>,
-        p: na::Matrix4<f64>,
-        u: f64,
-        fx: fn(na::Vector4<f64>, f64) -> na::Vector4<f64>,
-    ) -> na::SMatrix<f64, 4, 9> {
+        x: na::Vector4<f32>,
+        p: na::Matrix4<f32>,
+        u: f32,
+        fx: fn(na::Vector4<f32>, f32) -> na::Vector4<f32>,
+    ) -> na::SMatrix<f32, 4, 9> {
         let mut sigma_f = Self::sigma_points(&x, &p);
         for i in 0..9 {
             sigma_f.set_column(i, &fx(sigma_f.column(i).into_owned(), u));
@@ -81,19 +81,19 @@ impl UnscentedKalmanFilter {
     }
 
     // 推定した状態を返す
-    pub fn state(&self) -> na::Vector4<f64> {
+    pub fn state(&self) -> na::Vector4<f32> {
         self.x
     }
 
     fn unscented_transform<const S: usize>(
-        sigmas: &na::SMatrix<f64, S, 9>,
-        wm: &na::SVector<f64, 9>,
-        wc: &na::SVector<f64, 9>,
-        cov: &na::SMatrix<f64, S, S>,
-    ) -> (na::SVector<f64, S>, na::SMatrix<f64, S, S>) {
+        sigmas: &na::SMatrix<f32, S, 9>,
+        wm: &na::SVector<f32, 9>,
+        wc: &na::SVector<f32, 9>,
+        cov: &na::SMatrix<f32, S, S>,
+    ) -> (na::SVector<f32, S>, na::SMatrix<f32, S, S>) {
         let x = sigmas * wm;
-        let y = sigmas - na::SMatrix::<f64, S, 9>::from_columns(&[x, x, x, x, x, x, x, x, x]);
-        let mut tmp = na::SMatrix::<f64, S, S>::zeros();
+        let y = sigmas - na::SMatrix::<f32, S, 9>::from_columns(&[x, x, x, x, x, x, x, x, x]);
+        let mut tmp = na::SMatrix::<f32, S, S>::zeros();
         for i in 0..9 {
             tmp += wc[i] * y.column(i) * y.column(i).transpose();
         }
@@ -101,17 +101,17 @@ impl UnscentedKalmanFilter {
         (x, p)
     }
 
-    fn sigma_weight() -> (na::SVector<f64, 9>, na::SVector<f64, 9>) {
-        let mut wm = na::SVector::<f64, 9>::from_element(1.0 / (2.0 * Self::C));
-        let mut wc = na::SVector::<f64, 9>::from_element(1.0 / (2.0 * Self::C));
+    fn sigma_weight() -> (na::SVector<f32, 9>, na::SVector<f32, 9>) {
+        let mut wm = na::SVector::<f32, 9>::from_element(1.0 / (2.0 * Self::C));
+        let mut wc = na::SVector::<f32, 9>::from_element(1.0 / (2.0 * Self::C));
         wm[0] = Self::LAMBDA / Self::C;
         wc[0] = Self::LAMBDA / Self::C + 1.0 - Self::ALPHA.powi(2) + Self::BETA;
         (wm, wc)
     }
 
-    fn sigma_points(x: &na::Vector4<f64>, p: &na::Matrix4<f64>) -> na::SMatrix<f64, 4, 9> {
+    fn sigma_points(x: &na::Vector4<f32>, p: &na::Matrix4<f32>) -> na::SMatrix<f32, 4, 9> {
         let l = (Self::C * p).cholesky().expect("Cholesky fail").l();
-        na::SMatrix::<f64, 4, 9>::from_columns(&[
+        na::SMatrix::<f32, 4, 9>::from_columns(&[
             *x,
             *x + l.column(0),
             *x - l.column(0),
